Changes backend to node
This is a large change that involves changing the backend to an express/node
server, as well as moving of assets into their appropriate directories. The
metrics returned by the server are in "first pass" form: mostly opening the
same files as the original version and parsing the results. A lot of these
can be found using the native `os` module, which it will eventually be
switched over to.

Script assets are still intact, but will be fetched from npm in a future
change (jquery, bootstrap, and highcharts in particular). The minimized form
has been removed for now as there isn't a build step in place.